:zap: SV-487 Separated import methods in its own service
